Permito correr indexación en python 3
