+++ .gitignore
@@ -1,7 +1,7 @@
 
 .idea/*
 .vscode/*
-conf/*.sqlite
+conf/*.sqlite*
 *.pyc
 wks*.*
 conf/settings/local.py
