+++ django_datajsonar/apps/management/tasks.py
@@ -7,7 +7,8 @@ from django.utils import timezone
 from django_rq import job
 
 from django_datajsonar.apps.management.actions import DatasetIndexableToggler
-from django_datajsonar.apps.management.models import Node, DatasetIndexingFile, NodeRegisterFile
+from django_datajsonar.apps.management.models import Node, DatasetIndexingFile, NodeRegisterFile, \
+    ReadDataJsonTask
 from django_datajsonar.apps.management.strings import FILE_READ_ERROR
 from django_datajsonar.libs.indexing.catalog_reader import index_catalog
 
@@ -73,3 +74,16 @@ def process_node_register_file(register_file_id):
 
     register_file.state = NodeRegisterFile.PROCESSED
     register_file.save()
+
+
+def schedule_new_read_datajson_task():
+    try:
+        task = ReadDataJsonTask.objects.last()
+        if task and task.status in [ReadDataJsonTask.INDEXING, ReadDataJsonTask.RUNNING]:
+            return
+    except ReadDataJsonTask.DoesNotExist:
+        pass
+
+    new_task = ReadDataJsonTask()
+    new_task.save()
+    read_datajson.delay(new_task)
