+++ django_datajsonar/libs/indexing/tasks.py
@@ -69,13 +69,3 @@ def close_read_datajson_task():
         task.save()
 
 
-def schedule_new_read_datajson_task():
-    try:
-        task = ReadDataJsonTask.objects.last()
-        if task.status in [ReadDataJsonTask.INDEXING, ReadDataJsonTask.RUNNING]:
-            return
-    except ReadDataJsonTask.DoesNotExist:
-        pass
-
-    new_task = ReadDataJsonTask()
-    new_task.save()
